Finalizacion Backend

+++ src/main/java/com/juanma/portlofio/repository/RFormacion.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RFormacion extends JpaRepository<Formacion, Long> {
+public interface RFormacion extends JpaRepository<Formacion, Integer> {
 
 }
